# v2/openapi/itips-v2.yaml
openapi: 3.1.0
info:
  title: ITIPS V2 — Jetson Edge API (Dahua-native)
  version: 2.1.0
  description: |
    Two HTTP surfaces run on every Jetson:

    * **Public 5050** — operator dashboard, MJPEG feeds, SSE alert stream,
      worker-enrollment / plate-list management, Test Console.
    * **Inbound 8443** — backend → Jetson commands (B1–B5). Bearer auth.

    All AI runs on the Dahua cameras. The Jetson subscribes to camera
    events, packages signed evidence, and proxies operator clicks onto
    `faceRecognitionServer.cgi`, `recordUpdater.cgi`, `coaxialControlIO.cgi`,
    and `ptz.cgi`.
servers:
  - url: http://localhost:5050
    description: Public surface
  - url: http://localhost:8443
    description: Inbound surface (bearer token required)
...
security: []
paths:
...
  /api/snapshot/{camera_id}:
    get:
      tags: [Snapshots]
      summary: One-shot JPEG (from frame bus, falling back to the camera)
      parameters:
        - { name: camera_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: JPEG
          content:
            image/jpeg: {}
        "404": { description: unknown camera or no frame available }
        "503": { description: camera unreachable }
...
  /api/workers:
    get:
      tags: [Workers]
      summary: Local person_id → per-camera UID map
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkerList" }
    post:
      tags: [Workers]
      summary: Enroll a worker on every camera
      description: |
        Sends the JPEG to each camera's
        `/cgi-bin/faceRecognitionServer.cgi?action=addPerson`. The Jetson stores
        the resulting camera UIDs alongside the operator-supplied `person_id` so
        a later DELETE can find them again.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [full_name, image]
              properties:
                full_name: { type: string }
                person_id: { type: string, description: "Optional. Auto-generated if omitted." }
                sex: { type: string, enum: [Male, Female, Unknown] }
                image:
                  type: string
                  format: binary
                  description: JPEG or PNG.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkerEnrollResult" }
  /api/workers/{camera_id}:
    get:
      tags: [Workers]
      summary: Reality-check the workers group as the camera sees it
      parameters:
        - { name: camera_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  camera_id: { type: integer }
                  group_id: { type: string }
                  people:
                    type: array
                    items: { $ref: "#/components/schemas/CameraPerson" }
        "404": { description: camera not registered }
  /api/workers/{person_id}:
    delete:
      tags: [Workers]
      summary: Remove a worker from every camera that has them enrolled
      parameters:
        - { name: person_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  cameras_cleared:
                    type: array
                    items: { type: integer }
                  failures:
                    type: array
                    items: { type: string }
        "404": { description: unknown person_id }
  /api/plates:
    get:
      tags: [Plates]
      summary: List the plates in a camera-side list
      parameters:
        - { name: list_type, in: query, required: false,
            schema: { type: string, enum: [TrafficRedList, TrafficBlackList], default: TrafficRedList } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  list_type: { type: string }
                  cameras:
                    type: object
                    additionalProperties:
                      type: array
                      items: { $ref: "#/components/schemas/PlateRecord" }
    post:
      tags: [Plates]
      summary: Add a plate to one or more cameras
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlateAdd" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/plates/{camera_id}/{rec_no}:
    delete:
      tags: [Plates]
      summary: Remove a plate from one camera by its rec_no
      parameters:
        - { name: camera_id, in: path, required: true, schema: { type: integer } }
        - { name: rec_no,    in: path, required: true, schema: { type: integer } }
        - { name: list_type, in: query, required: false,
            schema: { type: string, enum: [TrafficRedList, TrafficBlackList], default: TrafficRedList } }
      responses:
        "200": { description: OK }
        "404": { description: unknown camera }
  /api/deterrence/{camera_id}/fire:
    post:
      tags: [Deterrence]
      summary: Engage the TiOC strobe and/or speaker
      parameters:
        - { name: camera_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                light: { type: boolean, default: true }
                speaker: { type: boolean, default: true }
      responses:
        "200": { description: OK }
        "404": { description: unknown camera }
        "502": { description: camera rejected the command (often non-TiOC model) }
  /api/deterrence/{camera_id}/standdown:
    post:
      tags: [Deterrence]
      summary: Silence the strobe + speaker and hand control back to camera AI
      parameters:
        - { name: camera_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: OK }
...
  /api/incidents/{incident_id}:
    get:
      tags: [Incidents]
      summary: Detailed view — metadata, manifest, signature, file list
      parameters:
        - { name: incident_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /api/incidents/{incident_id}/files/{filename}:
    get:
      tags: [Incidents]
      summary: Download a single file out of an incident package
      parameters:
        - { name: incident_id, in: path, required: true, schema: { type: string } }
        - { name: filename, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/octet-stream: {}
  /api/test/events/recent:
    get:
      tags: [TestConsole]
      summary: Last N raw events received from every camera
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/RawEvent" }
...
  /api/test/ptz/{camera_id}/{direction}:
    post:
      tags: [TestConsole]
      summary: Jog PTZ in a direction (start/stop pair)
      parameters:
        - { name: camera_id, in: path, required: true, schema: { type: integer } }
        - name: direction
          in: path
          required: true
          schema:
            type: string
            enum: [up, down, left, right, left_up, right_up, left_down, right_down,
                   zoom_wide, zoom_tele, focus_near, focus_far]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action: { type: string, enum: [start, stop], default: start }
                speed: { type: integer, minimum: 1, maximum: 8, default: 4 }
      responses:
        "200": { description: OK }
        "400": { description: unknown direction }
  /api/test/simulate/{event_type}:
    post:
      tags: [TestConsole]
      summary: Inject a synthetic event into the AlertEngine
      parameters:
        - name: event_type
          in: path
          required: true
          schema:
            type: string
            enum: [face_intruder, face_known, line_crossing, intrusion,
                   loitering, plate, fire, smoke]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                camera_id: { type: integer, default: 1 }
                plate_number: { type: string }
                similarity: { type: integer }
                bbox:
                  type: array
                  minItems: 4
                  maxItems: 4
                  items: { type: number }
      responses:
        "200": { description: OK }
  /api/test/inbound/{endpoint}:
    post:
      tags: [TestConsole]
      summary: Proxy to the inbound (8443) API with the configured bearer
      parameters:
        - name: endpoint
          in: path
          required: true
          schema:
            type: string
            enum: ["personnel/sync", "config", "maintenance/window", "commands", "firmware/update"]
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: relayed response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  status: { type: integer }
                  response: { description: body returned by the inbound API }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    WorkerList:
      type: object
      properties:
        workers:
          type: array
          items:
            type: object
            properties:
              person_id: { type: string }
              full_name: { type: string }
              cameras:
                type: object
                additionalProperties: { type: string }
        available_cameras:
          type: array
          items: { type: integer }
    WorkerEnrollResult:
      type: object
      properties:
        ok: { type: boolean }
        person_id: { type: string }
        cameras:
          type: object
          additionalProperties:
            type: string
            description: per-camera UID assigned by the camera
        failures:
          type: array
          items: { type: string }
    CameraPerson:
      type: object
      properties:
        uid: { type: string }
        group_id: { type: string }
        name: { type: string }
        sex: { type: string, nullable: true }
        has_image: { type: boolean }
    PlateRecord:
      type: object
      properties:
        rec_no: { type: integer }
        plate_number: { type: string }
        list_type: { type: string }
        master_of_car: { type: string, nullable: true }
        plate_color: { type: string, nullable: true }
        vehicle_color: { type: string, nullable: true }
...
    RawEvent:
      type: object
      properties:
        seq: { type: integer }
        camera_id: { type: integer }
        code: { type: string }
        action: { type: string, enum: [Start, Stop, Pulse] }
        index: { type: integer }
        data: { type: object, additionalProperties: true }
        has_jpeg: { type: boolean }
        ts: { type: number }
...